+++ Source/Android/app/build.gradle
@@ -112,8 +112,8 @@ dependencies {
     // For REST calls
     implementation 'com.android.volley:volley:1.2.1'
 
-    // For loading huge screenshots from the disk.
-    implementation 'com.squareup.picasso:picasso:2.71828'
+    // For loading game covers from disk and GameTDB
+    implementation 'com.github.bumptech.glide:glide:4.13.1'
 
     implementation 'com.nononsenseapps:filepicker:4.2.1'
 }
